requirements.txt

fastapi
uvicorn
python-dotenv
langchain==0.3.27 
langchain-community==0.3.29 
langchain-core==0.3.79 
langchain-pinecone==0.2.12 
langchain-groq==0.3.7 
langchain-huggingface==0.3.1 
langchain-text-splitters==0.3.10 
pinecone==7.3.0 
pinecone-plugin-assistant==1.8.0 
pinecone-plugin-interface==0.0.7 
pinecone-text==0.11.0 
sentence-transformers
pydantic==2.11.7 
pydantic-settings==2.10.1 
pydantic_core==2.33.2 
pypdf==6.0.0